No need for separate common/store/cluster.go
